+++ MainView.js
@@ -1,0 +1,52 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import EvilIcon from 'react-native-vector-icons/EvilIcons';
+import { createIconSet } from 'react-native-vector-icons';
+import glyphMap from './icons/MiKaIcons.json';
+
+const MiKaIcon = createIconSet(glyphMap, 'MiKaIcons', require('./icons/MiKaIcons.ttf'));
+
+export default class MainView extends React.Component {
+  render() {
+    return (
+      <View style={styles.appContainer}>
+        <View style={styles.configContainer}>
+          <EvilIcon name='play' size={85} color="#77CC7B" />
+          <EvilIcon name='gear' size={83} color="#77CC7B" />
+          <MiKaIcon name='pause' size={83} color="#77CC7B" />
+        </View>
+        <View style={styles.mainContainer}>
+          <Text style={styles.mainText}>Open up App.js to start working on your app!</Text>
+          <Text>Changes you make will automatically reload.</Text>
+          <Text>Shake your phone tasdfasdfo open the developer menu.</Text>
+        </View>
+      </View>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  appContainer: {
+    flex: 1,
+    flexDirection: 'row',
+    backgroundColor: '#0A6810',
+    justifyContent: 'space-between',
+    alignItems: 'stretch'
+  },
+
+  mainContainer: {
+
+  },
+
+  configContainer: {
+    width: 90,
+    backgroundColor: '#0E8A17',
+    alignItems: 'center',
+  },
+
+
+  mainText: {
+    color: '#77CC7B',
+  }
+
+});
