+++ README.md
@@ -1,3 +1,24 @@
+## Table of Contents
+
+* [pokerTimer Docu](#pokerTimer-docu)
+  * [View structure](#view-structure)
+* [Create React Native App Docu](#create-react-native-app-docu)
+
+# pokerTimer Docu
+
+## View structure
+
+```
+mainView
+  buttonBarView
+  contentView
+    timeView
+    blindView
+configView
+```
+
+# Create React Native App Docu
+
 This project was bootstrapped with [Create React Native App](https://github.com/react-community/create-react-native-app).
 
 Below you'll find information about performing common tasks. The most recent version of this guide is available [here](https://github.com/react-community/create-react-native-app/blob/master/react-native-scripts/template/README.md).
